py-experiments/gost-group-job/pyproject.toml
[tool.poetry]
name = "my-job"
version = "0.1.0"
description = ""
authors = ["Ramil Gadzhiev <rami.gaggi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.1"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
